requirements.txt
numpy==1.26.4
fastapi
transformers
openai-whisper
torch
uvicorn
python-dotenv